requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
python-dotenv==1.0.1
yarl==1.11.1
